agents: fix bugs concerning probe handling

+++ TLAgent/ULAgent.h
@@ -64,7 +64,7 @@ namespace tl_agent {
         void update_signal();
         bool do_get(paddr_t address);
         bool do_putfulldata(paddr_t address, uint8_t data[]);
-        bool do_putpartialdata(uint16_t address, uint8_t size, uint8_t mask, uint8_t data[]);
+        bool do_putpartialdata(uint16_t address, uint8_t size, uint32_t mask, uint8_t data[]);
     };
 
 }
